elevator/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
sqlparse==0.4.3
pytz==2022.7.1
psycopg2==2.9.5
tzdata==2022.7